mqtt-s3-glue-sns/requirements.txt
aws-cdk-lib==2.22.0
constructs<=11.0.0

aws_cdk.aws_iot_alpha
aws_cdk.aws_iot_actions_alpha

jsonpickle~=2.2.0
boto3~=1.24.1